+++ src/test/java/uk/gov/digital/ho/proving/income/api/IncomeThresholdCalculatorIT.java
@@ -12,12 +12,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest(
-    classes = IncomeThresholdCalculatorNew.class,
+    classes = IncomeThresholdCalculator.class,
     properties = {"threshold.yearly.base=1", "threshold.yearly.oneDependant=2", "threshold.yearly.remainingDependantsIncrement=3"})
-public class IncomeThresholdCalculatorNewIT {
+public class IncomeThresholdCalculatorIT {
 
     @Autowired
-    private IncomeThresholdCalculatorNew incomeThresholdCalculatorNew;
+    private IncomeThresholdCalculator incomeThresholdCalculator;
 
     private static final BigDecimal EXPECTED_YEARLY_ZERO_DEPENDANTS = BigDecimal.valueOf(1);
     private static final BigDecimal EXPECTED_YEARLY_ONE_DEPENDANTS = BigDecimal.valueOf(2);
@@ -25,16 +25,16 @@ public class IncomeThresholdCalculatorNewIT {
 
     @Test
     public void autowiring_baseThreshold_wiredFromProperties() {
-        assertThat(incomeThresholdCalculatorNew.yearlyThreshold(0)).isEqualTo(EXPECTED_YEARLY_ZERO_DEPENDANTS);
+        assertThat(incomeThresholdCalculator.yearlyThreshold(0)).isEqualTo(EXPECTED_YEARLY_ZERO_DEPENDANTS);
     }
 
     @Test
     public void autowiring_oneDependantThreshold_wiredFromProperties() {
-        assertThat(incomeThresholdCalculatorNew.yearlyThreshold(1)).isEqualTo(EXPECTED_YEARLY_ONE_DEPENDANTS);
+        assertThat(incomeThresholdCalculator.yearlyThreshold(1)).isEqualTo(EXPECTED_YEARLY_ONE_DEPENDANTS);
     }
 
     @Test
     public void autowiring_remainingDependantIncrememnt_wiredFromProperties() {
-        assertThat(incomeThresholdCalculatorNew.yearlyThreshold(2)).isEqualTo(EXPECTED_YEARLY_TWO_DEPENDANTS);
+        assertThat(incomeThresholdCalculator.yearlyThreshold(2)).isEqualTo(EXPECTED_YEARLY_TWO_DEPENDANTS);
     }
 }
